requirements.txt
google-genai>=1.16.0
chromadb>=0.5.0
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
tenacity>=8.3.0
rich>=13.7.0
typer>=0.12.0
httpx>=0.27.0
